+++ assets/javascripts/kitten/components/organisms/header-nav/stories/kisskissbankbank.js
@@ -1,17 +1,28 @@
 import React, { useState } from 'react'
+import { action } from '@storybook/addon-actions'
 import HeaderNav from '../index'
 import {
-  LendopolisLogo,
   KissKissBankBankLogo,
   HeaderMenu,
-  SearchIcon,
+  EnvelopeIcon,
+  EditIcon,
+  GiftIcon,
   COLORS,
   Text,
   AvatarWithTextAndBadge,
   domElementHelper,
   useWindowWidth,
   useDeepCompareEffect,
-} from '../../../..'
+  Button,
+  ButtonWithTooltip,
+  Overlay,
+  FlexWrapper,
+  HorizontalCrowdfundingCard,
+  ClockCircleIcon,
+  CheckedCircleIcon,
+  CLOSE_OVERLAY_EVENT,
+  OPEN_OVERLAY_EVENT,
+} from 'kitten'
 
 const HEADER_NAV_ID = 'kkbbAndCoHeaderNav'
 const getElementById = id => document.getElementById(id)
@@ -21,25 +32,22 @@ const getComputedLeft = id =>
   domElementHelper.getComputedLeft(getElementById(id))
 
 const InnerBurgerMenu = () => (
-  <HeaderMenu borderSide="right" largeItem noBorder>
+  <HeaderMenu borderSide="right" noBorder>
     <HeaderNav.Hidden min="l">
-      <HeaderMenu.Item href="#">Lancez votre projet</HeaderMenu.Item>
-      <HeaderMenu.Item href="#">Découvrez les projets</HeaderMenu.Item>
+      <HeaderMenu.Item href="#" isSelected>
+        Lancer un projet
+      </HeaderMenu.Item>
+      <HeaderMenu.Item href="#">Découvrir les projets</HeaderMenu.Item>
+      <HeaderMenu.Item href="#">Carte cadeau</HeaderMenu.Item>
+      <HeaderMenu.Item href="#">Vos questions</HeaderMenu.Item>
+      <HeaderMenu.Item href="#">Blog</HeaderMenu.Item>
+      <HeaderMenu.Item href="#" className="k-u-background-color-line1">
+        Se déconnecter
+      </HeaderMenu.Item>
+      <HeaderMenu.Item href="#" className="k-u-background-color-primary5">
+        Se connecter
+      </HeaderMenu.Item>
     </HeaderNav.Hidden>
-
-    <HeaderMenu.Item href="#" isSelected>
-      Vos questions
-    </HeaderMenu.Item>
-    <HeaderMenu.Item href="#">Blog</HeaderMenu.Item>
-    <HeaderMenu.Item external href="#">
-      <LendopolisLogo primaryColor="#000" />
-    </HeaderMenu.Item>
-    <HeaderMenu.Item external href="#">
-      <LendopolisLogo primaryColor="#000" />
-    </HeaderMenu.Item>
-    <HeaderMenu.Item external href="#">
-      <LendopolisLogo primaryColor="#000" />
-    </HeaderMenu.Item>
   </HeaderMenu>
 )
 
@@ -51,41 +59,167 @@ const InnerUserMenu = () => (
       hover: COLORS.background1,
     }}
   >
-    <HeaderMenu.Item href="#">Mon espace personnel</HeaderMenu.Item>
+    <HeaderMenu.Item href="#">Mon profil</HeaderMenu.Item>
     <HeaderMenu.Item href="#">Mes projets</HeaderMenu.Item>
-    <HeaderMenu.Item size="tiny" href="#">
-      Gluten Mag
-    </HeaderMenu.Item>
-    <HeaderMenu.Item size="tiny" href="#">
-      Kallix — Mobilier de bureau pour le télétravail
-    </HeaderMenu.Item>
-    <HeaderMenu.Item size="tiny" href="#">
-      Mon premier court-métrage
-    </HeaderMenu.Item>
     <HeaderMenu.Item href="#">Mes contributions</HeaderMenu.Item>
-    <HeaderMenu.Item href="#">Mes messages</HeaderMenu.Item>
     <HeaderMenu.Item href="#">Modifier mon profil</HeaderMenu.Item>
-    <HeaderMenu.Item button modifier="helium" href="#">
-      Mon projet en cours
+    <HeaderMenu.Item href="#" className="k-u-background-color-line1">
+      Se déconnecter
     </HeaderMenu.Item>
-    <HeaderMenu.Item href="#" className="k-u-background-color-background3">
-      Déconnexion
+    <HeaderMenu.Item href="#" className="k-u-background-color-primary5">
+      Se connecter
     </HeaderMenu.Item>
   </HeaderMenu>
 )
 
 const Navigation = () => (
   <>
-    <HeaderNav.Nav.Item href="#">Lancez votre projet</HeaderNav.Nav.Item>
-    <HeaderNav.Nav.Item href="#" className="is-selected">
-      Découvrez les projets
+    <HeaderNav.Nav.Item href="#" smallPadding>
+      Découvrir les projets
     </HeaderNav.Nav.Item>
+    <HeaderNav.Logged>
+      <HeaderNav.Nav.Item href="#" smallPadding>
+        Lancer un projet
+      </HeaderNav.Nav.Item>
+    </HeaderNav.Logged>
   </>
 )
 
-export const KissKissBankBankHeaderNavStory = ({ isLogged, args }) => {
+const SearchInput = () => {
+  const handleToggle = event => {
+    action('handleToggle')(event)
+    window.dispatchEvent(
+      new Event(
+        event.isDropdownExpanded ? OPEN_OVERLAY_EVENT : CLOSE_OVERLAY_EVENT,
+      ),
+    )
+  }
+
+  const handleInputChange = event => {
+    action('handleInputChange')(event)
+  }
+
+  return (
+    <HeaderNav.SearchInput
+      onMenuToggle={handleToggle}
+      searchInputProps={{
+        placeholder: 'Rechercher un projet…',
+        'aria-label': 'Rechercher un projet',
+        onChange: handleInputChange,
+      }}
+      searchButtonProps={{
+        type: 'submit',
+        'aria-label': 'Voir les résultats de la recherche',
+      }}
+      action="#form-action#"
+      method="POST"
+    >
+      <p className="k-u-a11y-visuallyHidden">4 résultats ont été trouvés</p>
+      <FlexWrapper gap={5} padding={15}>
+        <HorizontalCrowdfundingCard
+          title="Vivamus sagittis lacus vel augue laoreet rutrum faucibus dolor auctor."
+          description="Maecenas sed diam eget risus varius blandit sit amet non magna."
+          progress={37}
+          progressColor={COLORS.primary1}
+          info={
+            <div className="k-u-flex k-u-flex-alignItems-center">
+              <ClockCircleIcon
+                width="12"
+                height="12"
+                color={COLORS.background1}
+                bgColor={COLORS.primary1}
+                className="k-u-margin-right-noneHalf"
+              />
+              <Text weight="regular" size="micro" color="primary1">
+                Dernier jour&nbsp;!
+              </Text>
+            </div>
+          }
+        />
+        <HorizontalCrowdfundingCard
+          title="Morbi leo risus, porta ac consectetur ac, vestibulum at eros."
+          description="Aenean eu leo quam. Pellentesque ornare sem lacinia quam venenatis vestibulum."
+          progress={100}
+          progressColor={COLORS.valid}
+          info={
+            <div className="k-u-flex k-u-flex-alignItems-center">
+              <CheckedCircleIcon
+                width="12"
+                height="12"
+                color={COLORS.background1}
+                bgColor={COLORS.valid}
+                className="k-u-margin-right-noneHalf"
+              />
+              <Text weight="regular" size="micro" color="font1">
+                Réussi
+              </Text>
+            </div>
+          }
+        />
+        <HorizontalCrowdfundingCard
+          title="Aenean eu leo quam. Pellentesque ornare sem lacinia quam venenatis vestibulum."
+          description="Morbi leo risus, porta ac consectetur ac, vestibulum at eros."
+          progress={21}
+          progressColor={COLORS.primary1}
+          info={
+            <div className="k-u-flex k-u-flex-alignItems-center">
+              <ClockCircleIcon
+                width="12"
+                height="12"
+                color={COLORS.background1}
+                bgColor={COLORS.primary1}
+                className="k-u-margin-right-noneHalf"
+              />
+              <Text weight="regular" size="micro">
+                En cours
+              </Text>
+            </div>
+          }
+        />
+        <HorizontalCrowdfundingCard
+          title="Maecenas sed diam eget risus varius blandit sit amet non magna."
+          description="Vivamus sagittis lacus vel augue laoreet rutrum faucibus dolor auctor."
+          progress={60}
+          progressColor={COLORS.primary1}
+          info={
+            <div className="k-u-flex k-u-flex-alignItems-center">
+              <ClockCircleIcon
+                width="12"
+                height="12"
+                color={COLORS.background1}
+                bgColor={COLORS.primary1}
+                className="k-u-margin-right-noneHalf"
+              />
+              <Text weight="regular" size="micro">
+                En cours
+              </Text>
+            </div>
+          }
+        />
+      </FlexWrapper>
+
+      <Button
+        size="tiny"
+        type="submit"
+        fit="fluid"
+        className="k-u-margin-bottom-double k-u-margin-horizontal-double"
+        style={{
+          width: 'calc(100% - 40px)',
+        }}
+      >
+        Voir les résultats de la recherche
+      </Button>
+    </HeaderNav.SearchInput>
+  )
+}
+
+export const KissKissBankBankHeaderNavStory = ({
+  isLogged,
+  size,
+  hasMentorEditButton,
+  ...args
+}) => {
   const [burgerMenuWidth, setBurgerMenuWidth] = useState(null)
-  const [userMenuWidth, setUserMenuWidth] = useState(null)
   const windowWidth = useWindowWidth()
 
   useDeepCompareEffect(() => {
@@ -94,99 +228,139 @@ export const KissKissBankBankHeaderNavStory = ({ isLogged, args }) => {
         getComputedWidthElement(`${HEADER_NAV_ID}PlateformMenu`) +
         getComputedWidthElement(`${HEADER_NAV_ID}Logo`),
     )
-
-    if (isLogged) {
-      setTimeout(() => {
-        setUserMenuWidth(
-          getComputedWidthElement(`${HEADER_NAV_ID}UserMenu`) || '0',
-        )
-      }, 100)
-    }
   }, [isLogged, windowWidth])
 
   return (
-    <HeaderNav
-      id={HEADER_NAV_ID}
-      quickAccessProps={{
-        href: '#mainContent',
-        text: 'Aller au contenu principal',
-        zIndex: 300,
-      }}
-      isLogged={isLogged}
-      {...args}
-    >
-      <HeaderNav.BurgerMenu>
-        <InnerBurgerMenu dropdownContentWidth={burgerMenuWidth} />
-      </HeaderNav.BurgerMenu>
+    <>
+      <Overlay zIndex={1} position="fixed" />
 
-      <HeaderNav.Logo href="#">
-        <HeaderNav.Hidden min="xs">
+      <HeaderNav
+        id={HEADER_NAV_ID}
+        quickAccessProps={{
+          href: '#mainContent',
+          text: 'Aller au contenu principal',
+          zIndex: 300,
+        }}
+        isLogged={isLogged}
+        {...args}
+      >
+        <HeaderNav.BurgerMenu
+          dropdownContentWidth={burgerMenuWidth}
+          className="k-u-hidden@l-up"
+        >
+          <InnerBurgerMenu />
+        </HeaderNav.BurgerMenu>
+
+        <HeaderNav.Logo href="#logo">
           <KissKissBankBankLogo height="25" />
-        </HeaderNav.Hidden>
+        </HeaderNav.Logo>
 
-        <HeaderNav.Hidden max="xxs">
-          <KissKissBankBankLogo />
-        </HeaderNav.Hidden>
-      </HeaderNav.Logo>
-
-      <HeaderNav.Hidden max="m">
-        <HeaderNav.Nav>
+        <HeaderNav.Nav className="k-u-hidden@m-down">
           <Navigation />
         </HeaderNav.Nav>
-      </HeaderNav.Hidden>
 
-      <HeaderNav.Right>
-        <HeaderNav.Button
-          icon={<SearchIcon width="14" height="14" color={COLORS.font1} />}
-          backgroundColor={COLORS.background3}
-          backgroundColorHover={COLORS.line1}
-          color={COLORS.font1}
-          text="Rechercher"
-          href="#"
-          hiddenText={{ max: 'm' }}
-        />
+        <HeaderNav.Right>
+          <SearchInput />
 
-        <HeaderNav.Logged>
-          <HeaderNav.UserMenu dropdownContentWidth={userMenuWidth}>
-            <HeaderNav.UserMenu.Button>
-              <AvatarWithTextAndBadge>
-                <AvatarWithTextAndBadge.Image src="/kitten.jpg">
-                  <AvatarWithTextAndBadge.Badge>2</AvatarWithTextAndBadge.Badge>
-                </AvatarWithTextAndBadge.Image>
+          <ButtonWithTooltip
+            tooltipText="Carte cadeau"
+            buttonProps={{
+              as: 'a',
+              href: '#card',
+              icon: true,
+              size: 'tiny',
+              rounded: true,
+              modifier: 'hydrogen',
+            }}
+            className="k-u-hidden@m-down"
+          >
+            <GiftIcon width="15" height="16" />
+          </ButtonWithTooltip>
 
-                <AvatarWithTextAndBadge.Text
-                  textClassName="k-u-hidden@xs-down"
-                  withEllipsisOverflow={true}
-                >
-                  <Text lineHeight="normal" weight="regular">
-                    Jean Charles Édouard
-                  </Text>
-                </AvatarWithTextAndBadge.Text>
-              </AvatarWithTextAndBadge>
-            </HeaderNav.UserMenu.Button>
+          <HeaderNav.Logged>
+            <ButtonWithTooltip
+              tooltipText="Mes messages"
+              buttonProps={{
+                as: 'a',
+                href: '#hello',
+                icon: true,
+                size: 'tiny',
+                rounded: true,
+                modifier: 'hydrogen',
+              }}
+              className="k-u-hidden@m-down"
+              children={<EnvelopeIcon />}
+            />
 
-            <HeaderNav.UserMenu.Navigation>
-              <InnerUserMenu />
-            </HeaderNav.UserMenu.Navigation>
-          </HeaderNav.UserMenu>
-        </HeaderNav.Logged>
+            {hasMentorEditButton && (
+              <Button
+                as="a"
+                size="tiny"
+                modifier="helium"
+                borderRadius={100}
+                className="k-u-hidden@m-down"
+              >
+                <EditIcon />
+                <span>Éditer ma page</span>
+              </Button>
+            )}
 
-        <HeaderNav.LoggedOut>
-          <HeaderNav.Button
-            icon={
-              <HeaderNav.Hidden min="s">
-                <SearchIcon />
-              </HeaderNav.Hidden>
-            }
-            backgroundColor={COLORS.primary1}
-            backgroundColorHover={COLORS.primary2}
-            color={COLORS.background1}
-            text="Se connecter / S'inscrire"
-            href="#"
-            hiddenText={{ max: 'xs' }}
-          />
-        </HeaderNav.LoggedOut>
-      </HeaderNav.Right>
-    </HeaderNav>
+            <HeaderNav.UserMenu dropdownAnchorSide="right">
+              <HeaderNav.UserMenu.Button
+                hasArrow
+                backgroundColor={COLORS.background1}
+                backgroundColorHover="transparent"
+                backgroundColorActive="transparent"
+                color={COLORS.font1}
+                colorHover={COLORS.primary1}
+                colorActive={COLORS.font1}
+              >
+                <AvatarWithTextAndBadge>
+                  <AvatarWithTextAndBadge.Image src="/kitten.jpg">
+                    <AvatarWithTextAndBadge.Badge>
+                      2
+                    </AvatarWithTextAndBadge.Badge>
+                  </AvatarWithTextAndBadge.Image>
+
+                  <AvatarWithTextAndBadge.Text
+                    className="k-u-hidden@xs-down"
+                    withEllipsisOverflow={true}
+                  >
+                    <Text lineHeight="normal" weight="regular">
+                      Jean Charles Édouard
+                    </Text>
+                  </AvatarWithTextAndBadge.Text>
+                </AvatarWithTextAndBadge>
+              </HeaderNav.UserMenu.Button>
+              <HeaderNav.UserMenu.Navigation>
+                <InnerUserMenu />
+              </HeaderNav.UserMenu.Navigation>
+            </HeaderNav.UserMenu>
+          </HeaderNav.Logged>
+
+          <HeaderNav.LoggedOut>
+            <HeaderNav.Button
+              as="a"
+              backgroundColor={COLORS.background1}
+              backgroundColorHover={COLORS.background1}
+              color={COLORS.font1}
+              colorHover={COLORS.primary1}
+              text="Se connecter"
+              className="k-u-hidden@m-down"
+            />
+
+            <Button
+              as="a"
+              size="tiny"
+              modifier="helium"
+              borderRadius={100}
+              className="k-u-hidden@m-down"
+            >
+              Lancer un projet
+            </Button>
+          </HeaderNav.LoggedOut>
+        </HeaderNav.Right>
+      </HeaderNav>
+    </>
   )
 }